# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wanderbound",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/albums/pdf/download/{token}": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Download Pdf",
        "operationId": "download_pdf",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Read Album",
        "operationId": "read_album",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "albums"
        ],
        "summary": "Update Album",
        "operationId": "update_album",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlbumUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/media": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Read Media",
        "operationId": "read_media",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumMedia"
                  },
                  "title": "Response Read Media"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/steps": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Read Steps",
        "operationId": "read_steps",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Step"
                  },
                  "title": "Response Read Steps"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/segments": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Read Segments",
        "operationId": "read_segments",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SegmentOutline"
                  },
                  "title": "Response Read Segments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/segments/points": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Read Segment Points",
        "operationId": "read_segment_points",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          },
          {
            "name": "from_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "From Time"
            }
          },
          {
            "name": "to_time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "To Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Segment"
                  },
                  "title": "Response Read Segment Points"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/steps/{sid}": {
      "patch": {
        "tags": [
          "albums"
        ],
        "summary": "Update Step",
        "operationId": "update_step",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          },
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/segments/adjust-boundary": {
      "patch": {
        "tags": [
          "albums"
        ],
        "summary": "Adjust Segment Boundary",
        "operationId": "adjust_segment_boundary",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoundaryAdjust"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SegmentOutline"
                  },
                  "title": "Response Adjust Segment Boundary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/print-bundle": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Read Print Bundle",
        "operationId": "read_print_bundle",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/albums/{aid}/pdf/generate": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Generate Pdf",
        "operationId": "generate_pdf",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          },
          {
            "name": "dark",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Dark"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/PdfQueued"
                      },
                      {
                        "$ref": "#/components/schemas/PdfProgress"
                      },
                      {
                        "$ref": "#/components/schemas/PdfDone"
                      },
                      {
                        "$ref": "#/components/schemas/PdfError"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "type",
                      "mapping": {
                        "queued": "#/components/schemas/PdfQueued",
                        "progress": "#/components/schemas/PdfProgress",
                        "done": "#/components/schemas/PdfDone",
                        "error": "#/components/schemas/PdfError"
                      }
                    }
                  },
                  "title": "Response 200 Generate Pdf"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlbumMedia": {
        "properties": {
          "uid": {
            "type": "integer",
            "title": "Uid"
          },
          "aid": {
            "type": "string",
            "title": "Aid"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "storage_path": {
            "type": "string",
            "maxLength": 255,
            "title": "Storage Path"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "source_ref_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "aid",
          "name",
          "kind",
          "storage_path",
          "width",
          "height",
          "byte_size"
        ],
        "title": "AlbumMedia"
      },
      "AlbumMeta": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "maxLength": 255,
            "title": "Subtitle"
          },
          "hidden_steps": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Hidden Steps"
          },
          "hidden_headers": {
            "items": {
              "type": "string",
              "enum": [
                "cover-front",
                "cover-back",
                "overview",
                "full-map"
              ]
            },
            "type": "array",
            "title": "Hidden Headers"
          },
          "maps_ranges": {
            "items": {
              "$ref": "#/components/schemas/DateRange"
            },
            "type": "array",
            "title": "Maps Ranges"
          },
          "front_cover_photo": {
            "type": "string",
            "maxLength": 255,
            "title": "Front Cover Photo"
          },
          "back_cover_photo": {
            "type": "string",
            "maxLength": 255,
            "title": "Back Cover Photo"
          },
          "font": {
            "type": "string",
            "title": "Font",
            "default": "Assistant"
          },
          "body_font": {
            "type": "string",
            "title": "Body Font",
            "default": "Frank Ruhl Libre"
          },
          "safe_margin_mm": {
            "type": "integer",
            "title": "Safe Margin Mm",
            "default": 5
          },
          "media_resolution_warning_preset": {
            "type": "string",
            "enum": [
              "off",
              "relaxed",
              "print"
            ],
            "title": "Media Resolution Warning Preset",
            "default": "relaxed"
          },
          "uid": {
            "type": "integer",
            "title": "Uid"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "colors": {
            "patternProperties": {
              "[a-zA-Z]{2}|00": {
                "type": "string",
                "pattern": "#[0-9a-fA-F]{6}"
              }
            },
            "type": "object",
            "title": "Colors"
          }
        },
        "type": "object",
        "required": [
          "title",
          "subtitle",
          "front_cover_photo",
          "back_cover_photo",
          "uid",
          "id",
          "colors"
        ],
        "title": "AlbumMeta",
        "description": "GET/PATCH response - everything except media."
      },
      "AlbumUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "hidden_steps": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hidden Steps"
          },
          "hidden_headers": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "cover-front",
                    "cover-back",
                    "overview",
                    "full-map"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hidden Headers"
          },
          "maps_ranges": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DateRange"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maps Ranges"
          },
          "front_cover_photo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Front Cover Photo"
          },
          "back_cover_photo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Back Cover Photo"
          },
          "font": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Font",
            "default": "Assistant"
          },
          "body_font": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Font",
            "default": "Frank Ruhl Libre"
          },
          "safe_margin_mm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Margin Mm",
            "default": 5
          },
          "media_resolution_warning_preset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "off",
                  "relaxed",
                  "print"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Resolution Warning Preset",
            "default": "relaxed"
          }
        },
        "type": "object",
        "title": "AlbumUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SegmentOutline": {
        "properties": {
          "start_time": {
            "type": "number",
            "title": "Start Time"
          },
          "end_time": {
            "type": "number",
            "title": "End Time"
          },
          "kind": {
            "$ref": "#/components/schemas/SegmentKind"
          },
          "timezone_id": {
            "type": "string",
            "title": "Timezone Id"
          },
          "start_coord": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Start Coord"
          },
          "end_coord": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "End Coord"
          }
        },
        "type": "object",
        "required": [
          "start_time",
          "end_time",
          "kind",
          "timezone_id",
          "start_coord",
          "end_coord"
        ],
        "title": "SegmentOutline",
        "description": "Lightweight segment projection - no points or route."
      },
...
      "Step": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover"
          },
          "pages": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Pages"
          },
          "unused": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unused"
          },
          "uid": {
            "type": "integer",
            "title": "Uid"
          },
          "aid": {
            "type": "string",
            "title": "Aid"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "timezone_id": {
            "type": "string",
            "maxLength": 255,
            "title": "Timezone Id"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "elevation": {
            "type": "integer",
            "title": "Elevation"
          },
          "weather": {
            "$ref": "#/components/schemas/Weather"
          },
          "datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Datetime",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "pages",
          "unused",
          "uid",
          "aid",
          "id",
          "timestamp",
          "timezone_id",
          "location",
          "elevation",
          "weather",
          "datetime"
        ],
        "title": "Step"
      },
...
    }
  }
}